requirements.txt
click==8.1.3
cmake==3.25.2
dlib==19.24.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.4.0